pyproject.toml
[project]
name = "llm-poc"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "ipykernel>=6.29.5",
    "isort>=5.13.2",
    "nbstripout>=0.8.1",
    "openai>=1.59.9",
    "pre-commit>=4.1.0",
    "python-dotenv>=1.0.1",
    "ruff>=0.9.2",
]

[tool.ruff]
# Enable specific checks
## F401: Detects unused imports.
## I001: Enforces import sorting (similar to isort).
## I004: Flags duplicate imports.

# Automatically sort imports
[tool.ruff.lint]
fixable = ["I001"]
select = ["F401", "I001", "F403"]

# Exclude certain files/directories (if needed)
exclude = ["tests/*"]
